+++ .node-version
@@ -1,1 +1,1 @@
-18.16.1
+20.6.1
