+++ package.json
@@ -45,10 +45,10 @@
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-switch": "^1.0.3",
     "@radix-ui/react-toggle": "^1.0.3",
-    "@supabase/auth-helpers-nextjs": "^0.8.1",
-    "@supabase/supabase-js": "^2.38.2",
+    "@supabase/auth-helpers-nextjs": "^0.8.2",
+    "@supabase/supabase-js": "^2.38.4",
     "@vercel/analytics": "^1.1.1",
-    "ai": "^2.2.16",
+    "ai": "^2.2.20",
     "animate.css": "^4.1.1",
     "api": "^6.1.1",
     "body-parser": "^1.20.2",
@@ -56,10 +56,10 @@
     "clsx": "^2.0.0",
     "date-fns": "^2.30.0",
     "htmr": "^1.0.2",
-    "lucide-react": "^0.288.0",
+    "lucide-react": "^0.289.0",
     "microcms-js-sdk": "^2.7.0",
     "million": "^2.6.4",
-    "next": "13.5.6",
+    "next": "14.0.0",
     "next-seo": "^6.1.0",
     "openai-edge": "^1.2.2",
     "react": "18.2.0",
@@ -74,26 +74,26 @@
     "remark-gemoji": "^8.0.0",
     "remark-gfm": "^4.0.0",
     "sharp": "^0.32.6",
-    "swiper": "^10.3.1",
+    "swiper": "^11.0.3",
     "swr": "^2.2.4",
     "tailwind-merge": "^1.14.0",
-    "tailwind-variants": "^0.1.14",
+    "tailwind-variants": "^0.1.17",
     "tailwindcss-animate": "^1.0.7",
-    "the-new-css-reset": "^1.11.0",
+    "the-new-css-reset": "^1.11.1",
     "ts-pattern": "^5.0.5",
     "umt": "^1.1.7",
     "zod": "^3.22.4"
   },
   "devDependencies": {
     "@babel/core": "^7.23.2",
-    "@biomejs/biome": "1.3.0",
+    "@biomejs/biome": "1.3.1",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@kuma-ui/next-plugin": "^1.1.12",
-    "@markuplint/jsx-parser": "^3.10.0",
-    "@markuplint/react-spec": "^3.11.0",
-    "@mui/icons-material": "^5.14.14",
-    "@mui/material": "^5.14.14",
+    "@markuplint/jsx-parser": "^3.11.0",
+    "@markuplint/react-spec": "^3.12.0",
+    "@mui/icons-material": "^5.14.15",
+    "@mui/material": "^5.14.15",
     "@playwright/test": "^1.39.0",
     "@storybook/addon-actions": "^7.5.1",
     "@storybook/addon-essentials": "^7.5.1",
@@ -106,33 +106,33 @@
     "@storybook/nextjs": "^7.5.1",
     "@storybook/react": "^7.5.1",
     "@storybook/testing-library": "^0.2.2",
-    "@swc/core": "^1.3.93",
+    "@swc/core": "^1.3.95",
     "@swc/jest": "^0.2.29",
     "@testing-library/jest-dom": "^6.1.4",
     "@testing-library/react": "^14.0.0",
     "@types/jest": "^29.5.6",
-    "@types/node": "20.8.7",
+    "@types/node": "20.8.9",
     "@types/pg": "^8.10.7",
     "@types/prismjs": "^1.26.2",
-    "@types/react": "18.2.30",
+    "@types/react": "18.2.33",
     "@types/react-dom": "18.2.14",
     "autoprefixer": "^10.4.16",
     "babel-loader": "^9.1.3",
-    "eslint": "^8.51.0",
-    "eslint-config-next": "^13.5.6",
-    "eslint-plugin-import": "^2.28.1",
+    "eslint": "^8.52.0",
+    "eslint-config-next": "^14.0.0",
+    "eslint-plugin-import": "^2.29.0",
     "eslint-plugin-jest-dom": "^5.1.0",
     "eslint-plugin-storybook": "^0.6.15",
     "eslint-plugin-testing-library": "^6.1.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "markuplint": "^3.13.0",
+    "markuplint": "^3.14.0",
     "next-intercept-stdout": "^1.0.1",
     "next-router-mock": "^0.9.10",
     "postcss": "^8.4.31",
-    "sass": "^1.69.4",
+    "sass": "^1.69.5",
     "storybook": "^7.5.1",
-    "tailwindcss": "^3.3.3",
+    "tailwindcss": "^3.3.5",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "turbo": "^1.10.16",
