+++ src/app/index/useTechComponent.tsx
@@ -1,4 +1,3 @@
-"use client";
 import Card from "@mui/material/Card";
 import CardContent from "@mui/material/CardContent";
 import CardMedia from "@mui/material/CardMedia";
