+++ src/components/ui/auto-form/fields/number.tsx
@@ -1,0 +1,18 @@
+import { AutoFormInputComponentProps } from "../types";
+
+import AutoFormInput from "./input";
+
+export default function AutoFormNumber({
+  fieldProps,
+  ...props
+}: AutoFormInputComponentProps) {
+  return (
+    <AutoFormInput
+      fieldProps={{
+        type: "number",
+        ...fieldProps,
+      }}
+      {...props}
+    />
+  );
+}
